requirements.txt
alembic==1.13.1
aniso8601==9.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
blinker==1.8.1
cffi==1.16.0
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.3
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
referencing==0.35.1
rpds-py==0.18.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.3
zipp==3.18.1
